drop usage toolbar wrapper
